Create `dashboard` and `wp-dashboard` packages (#8646)
