+++ .storybook/main.cjs
@@ -17,7 +17,7 @@
 module.exports = {
   stories: [
     './stories/**/*.js',
-    '../assets/src/dashboard/**/stories/*.@(js|mdx)',
+    '../packages/dashboard/src/**/stories/*.@(js|mdx)',
     '../packages/story-editor/src/**/stories/*.@(js|mdx)',
     '../packages/activation-notice/src/**/stories/*.@(js|mdx)',
     '../packages/design-system/src/**/stories/*.@(js|mdx)',
