+++ package.json
@@ -32,44 +32,7 @@
       "packages/*"
     ]
   },
-  "dependencies": {
-    "@web-stories-wp/activation-notice": "*",
-    "@web-stories-wp/animation": "*",
-    "@web-stories-wp/date": "*",
-    "@web-stories-wp/design-system": "*",
-    "@web-stories-wp/fonts": "*",
-    "@web-stories-wp/glider": "*",
-    "@web-stories-wp/i18n": "*",
-    "@web-stories-wp/imgareaselect": "*",
-    "@web-stories-wp/jest-amp": "*",
-    "@web-stories-wp/jest-puppeteer-amp": "*",
-    "@web-stories-wp/karma-fixture": "*",
-    "@web-stories-wp/media": "*",
-    "@web-stories-wp/migration": "*",
-    "@web-stories-wp/patterns": "*",
-    "@web-stories-wp/react": "*",
-    "@web-stories-wp/stickers": "*",
-    "@web-stories-wp/stories-block": "*",
-    "@web-stories-wp/stories-carousel": "*",
-    "@web-stories-wp/stories-lightbox": "*",
-    "@web-stories-wp/story-editor": "*",
-    "@web-stories-wp/templates": "*",
-    "@web-stories-wp/text-sets": "*",
-    "@web-stories-wp/tinymce-button": "*",
-    "@web-stories-wp/tracking": "*",
-    "@web-stories-wp/units": "*",
-    "@web-stories-wp/widget": "*",
-    "@web-stories-wp/wp-story-editor": "*",
-    "@wordpress/api-fetch": "^5.2.1",
-    "flagged": "^2.0.1",
-    "history": "^5.0.0",
-    "prop-types": "^15.7.2",
-    "query-string": "^7.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "styled-components": "^5.3.0",
-    "uuid": "^8.3.2"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@ampproject/toolbox-optimizer": "^2.8.3",
     "@axe-core/puppeteer": "^4.2.2",
@@ -99,11 +62,15 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/react-hooks": "^7.0.1",
     "@testing-library/user-event": "^13.2.1",
+    "@web-stories-wp/dashboard": "*",
+    "@web-stories-wp/design-system": "*",
     "@web-stories-wp/e2e-tests": "*",
+    "@web-stories-wp/jest-amp": "*",
     "@web-stories-wp/karma-cuj-reporter": "*",
     "@web-stories-wp/karma-failed-tests-reporter": "*",
     "@web-stories-wp/karma-puppeteer-client": "*",
     "@web-stories-wp/karma-puppeteer-launcher": "*",
+    "@web-stories-wp/story-editor": "*",
     "@wordpress/babel-plugin-import-jsx-pragma": "^3.1.0",
     "@wordpress/dependency-extraction-webpack-plugin": "^3.2.1",
     "@wordpress/eslint-plugin": "^9.1.0",
@@ -166,10 +133,13 @@
     "patch-package": "^6.4.7",
     "prettier": "^2.3.2",
     "puppeteer": "^10.1.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-test-renderer": "^17.0.2",
     "rtlcss-webpack-plugin": "^4.0.6",
     "source-map-loader": "^1.1.3",
     "storybook-rtl-addon": "^0.3.3",
+    "styled-components": "^5.3.0",
     "stylelint": "^13.12.0",
     "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-recommended": "^5.0.0",
