+++ packages/dashboard/README.md
@@ -1,0 +1,3 @@
+# Dashboard
+
+Web stories core dashboard.
