+++ packages/dashboard/src/index.js
@@ -17,62 +17,10 @@
 /**
  * Internal dependencies
  */
-// The __webpack_public_path__ assignment will be done after the imports.
-// That's why the public path assignment is in its own dedicated module and imported here at the very top.
-// See https://webpack.js.org/guides/public-path/#on-the-fly
-import './publicPath';
-
-/**
- * External dependencies
- */
-import { setAppElement } from '@web-stories-wp/design-system';
-import { StrictMode } from 'react';
-import { render } from 'react-dom';
-import { FlagsProvider } from 'flagged';
-import { updateSettings } from '@web-stories-wp/date';
-import { initializeTracking } from '@web-stories-wp/tracking';
-
-/**
- * Internal dependencies
- */
 import App from './app';
-import './style.css'; // This way the general dashboard styles are loaded before all the component styles.
 
-/**
- * Initializes the Web Stories dashboard screen.
- *
- * @param {string} id       ID of the root element to render the screen in.
- * @param {Object} config   Story editor settings.
- * @param {Object} flags    The flags for the application.
- */
-const initialize = async (id, config, flags) => {
-  const appElement = document.getElementById(id);
-
-  // see http://reactcommunity.org/react-modal/accessibility/
-  setAppElement(appElement);
-
-  updateSettings(config.locale);
-
-  // Already tracking screen views in AppContent, no need to send page views as well.
-  await initializeTracking('Dashboard', false);
-
-  render(
-    <FlagsProvider features={flags}>
-      <StrictMode>
-        <App config={config} />
-      </StrictMode>
-    </FlagsProvider>,
-    appElement
-  );
-};
-
-const initializeWithConfig = () => {
-  const { id, config, flags } = window.webStoriesDashboardSettings;
-  initialize(id, config, flags);
-};
-
-if ('loading' === document.readyState) {
-  document.addEventListener('DOMContentLoaded', initializeWithConfig);
-} else {
-  initializeWithConfig();
-}
+export default App;
+export { GlobalStyle as DashboardGlobalStyle } from './theme';
+export { default as DashboardKeyboardOnlyOutline } from './utils/keyboardOnlyOutline';
+export { ConfigProvider as DashboardConfigProvider } from './app/config';
+export { default as ApiProvider } from './app/api/apiProvider';
