+++ packages/wp-dashboard/README.md
@@ -1,0 +1,3 @@
+# WP Dashboard
+
+Contains WordPress integration of dashboard.
