+++ packages/wp-dashboard/package.json
@@ -1,0 +1,23 @@
+{
+	"name": "@web-stories-wp/wp-dashboard",
+	"private": true,
+	"engines": {
+		"node": ">= 15",
+		"npm": "~7"
+	},
+	"exports": {
+		"default": "./src/index.js"
+	},
+	"main": "./src/index.js",
+	"type": "module",
+	"dependencies": {
+		"@web-stories-wp/date": "*",
+		"@web-stories-wp/dashboard": "*",
+		"@web-stories-wp/design-system": "*",
+		"@web-stories-wp/tracking": "*",
+		"flagged": "^2.0.1",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2"
+	},
+	"devDependencies": {}
+}
