+++ packages/wp-dashboard/src/publicPath.js
